+++ src/main/java/br/ufrn/pedrogalvao/atlas/sensor/SensorService.java
@@ -3,9 +3,11 @@ package br.ufrn.pedrogalvao.atlas.sensor;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import br.ufrn.pedrogalvao.atlas.exception.MissionNotFoundException;
 import br.ufrn.pedrogalvao.atlas.exception.SensorCreationNotAllowedException;
+import br.ufrn.pedrogalvao.atlas.exception.SensorDeletionNotAllowedException;
 import br.ufrn.pedrogalvao.atlas.exception.SensorNotFoundException;
 import br.ufrn.pedrogalvao.atlas.mission.Mission;
 import br.ufrn.pedrogalvao.atlas.mission.MissionRepository;
@@ -55,9 +57,16 @@ public class SensorService {
                 .orElseThrow(() -> new SensorNotFoundException(sensorId));
     }
 
-    public void delete(Long sensorId) {
+    @Transactional
+    public void delete(Long missionId, Long sensorId) {
+    	Mission mission = missionRepository.findById(missionId)
+                .orElseThrow(() -> new MissionNotFoundException(missionId));
+    	
     	sensorRepository.findById(sensorId)
         		.orElseThrow(() -> new SensorNotFoundException(sensorId));
+    	if (mission.getStatus() != MissionStatus.PLANNED) {
+        	throw new SensorDeletionNotAllowedException();
+        }
     	sensorRepository.deleteById(sensorId);
     }
 }
